refactor(kitchenpos) 기능 우선 패키지 구성하기
