+++ src/main/java/kitchenpos/menu/domain/MenuProduct.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.menu.domain;
+
+import kitchenpos.product.domain.Product;
 
 import javax.persistence.*;
 import java.util.UUID;
