+++ src/main/java/kitchenpos/menu/repository/JpaMenuRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.menu.repository;
 
 import kitchenpos.menu.domain.Menu;
 import org.springframework.data.jpa.repository.JpaRepository;
