+++ src/main/java/kitchenpos/order/domain/Order.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.order.domain;
+
+import kitchenpos.ordertable.domain.OrderTable;
 
 import javax.persistence.*;
 import java.time.LocalDateTime;
