+++ src/main/java/kitchenpos/order/repository/JpaOrderRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.order.repository;
 
+import kitchenpos.order.domain.Order;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
