+++ src/main/java/kitchenpos/ordertable/repository/JpaOrderTableRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.ordertable.repository;
 
+import kitchenpos.ordertable.domain.OrderTable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
