+++ src/main/java/kitchenpos/product/repository/JpaProductRepository.java
@@ -1,5 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.product.repository;
 
+import kitchenpos.product.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.UUID;
