+++ src/main/java/kitchenpos/product/repository/ProductRepository.java
@@ -1,4 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.product.repository;
+
+import kitchenpos.product.domain.Product;
 
 import java.util.List;
 import java.util.Optional;
